Add working example
* Use webpack-dev-server to serve the example code
* Update and tidy up the README
