+++ example/.eslintrc.js
@@ -1,0 +1,5 @@
+module.exports = {
+  "rules": {
+    "no-console": 0
+  }
+};
