+++ example/index.html
@@ -1,0 +1,11 @@
+<!DOCTYPE html>
+<html>
+<head>
+  <meta charset="utf-8">
+  <title>example | react-svg</title>
+</head>
+<body>
+  <div class="Root"></div>
+  <script src="bundle.js"></script>
+</body>
+</html
